\name{print.coxphw.predict}

\alias{print.coxphw.predict}

\title{ Print Method for Objects of Class \code{predict.coxphw} }

\description{ This class of objects is returned by the \code{predict.coxphw} function.
  Objects of this class have methods for the functions \code{print} and \code{plot}. }

\usage{ \method{print}{coxphw.predict}(x, ...) }

\arguments{
  \item{x}{ object of class \code{coxphw.predict}. }
  \item{\dots}{ further arguments. }
}

%%\details{}

%%\value{  ~Describe the value returned }

%% \references{ ~put references to the literature/web site here ~ }

\author{  Daniela Dunkler }

%% \note{  ~~further notes~~ }

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{ \code{\link{coxphw}}, \code{\link{predict.coxphw}} \code{\link{plot.coxphw.predict}} }

%%\examples{}

\keyword{ utilities }
